fix Dropdown item styles
